% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dated.R
\name{date_this}
\alias{date_this}
\title{Get Date}
\usage{
date_this(without_punct = FALSE)
}
\arguments{
\item{without_punct}{default is FALSE. If true, returns date in YYYYmmdd format. Otherwise YYYY-mm-dd format.}
}
\description{
Get Date
}
