% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_date_col.R
\name{qa_date_col}
\alias{qa_date_col}
\title{Returns dataframe of the outputs for all the parsing options for a date vector}
\usage{
qa_date_col(dataframe, date_col)
}
\description{
Returns dataframe of the outputs for all the parsing options for a date vector
}
