% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_timestamp_col.R
\name{qa_timestamp_col}
\alias{qa_timestamp_col}
\title{Returns a dataframe of the different parsing methods for a timestamp column}
\usage{
qa_timestamp_col(dataframe, timestamp_col)
}
\description{
Returns a dataframe of the different parsing methods for a timestamp column
}
