% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamped.R
\name{stamp_this}
\alias{stamp_this}
\title{Get Timestamp}
\usage{
stamp_this(without_punct = FALSE)
}
\arguments{
\item{without_punct}{defaults to FALSE. If true removes all punctuation so output is YYYYmmddHHMMSS format.}
}
\description{
Get Timestamp
}
